refactor: restructure Header component and separate MobileMenu functionality
- Removed the existing Header component and created a new Header component with improved structure.
- Integrated MobileMenu for better mobile navigation experience.
- Enhanced styling and animations for the header and buttons.
- Maintained existing functionality while improving code organization.
